+++ GroupProject/src/main/java/Bookstore/Model/Owner.java
@@ -1,4 +1,4 @@
-package Bookstore.repositories;
+package Bookstore.Model;
 
 import javax.persistence.*;
 import java.util.ArrayList;
