+++ GroupProject/src/main/java/Bookstore/repositories/BookRepository.java
@@ -1,6 +1,7 @@
 package Bookstore.repositories;
 
 
+import Bookstore.Model.Book;
 import org.springframework.data.repository.PagingAndSortingRepository;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
