+++ GroupProject/src/main/java/Bookstore/repositories/BookstoreRepository.java
@@ -1,5 +1,6 @@
 package Bookstore.repositories;
 
+import Bookstore.Model.Bookstore;
 import org.springframework.data.repository.PagingAndSortingRepository;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
