+++ GroupProject/src/main/java/Bookstore/repositories/CartRepository.java
@@ -1,6 +1,7 @@
 package Bookstore.repositories;
 
 
+import Bookstore.Model.Cart;
 import org.springframework.data.repository.PagingAndSortingRepository;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
