+++ GroupProject/src/main/java/Bookstore/repositories/UserRepository.java
@@ -1,6 +1,7 @@
 package Bookstore.repositories;
 
 
+import Bookstore.Model.User;
 import org.springframework.data.repository.PagingAndSortingRepository;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
 import org.springframework.stereotype.Repository;
